map_navigation_project/requirements.txt
# Core Habitat dependencies
# Note: habitat-sim and habitat-lab should be installed separately following official instructions

# Computer Vision and Image Processing
opencv-python>=4.5.0
matplotlib>=3.3.0
Pillow>=8.0.0

# Numerical Computing
numpy>=1.19.0
scipy>=1.7.0

# Quaternion operations (if not available through habitat-sim)
quaternion>=2022.4.0

# Configuration and Data Handling
PyYAML>=5.4.0
omegaconf>=2.1.0
hydra-core>=1.1.0

# Visualization and Plotting
seaborn>=0.11.0
plotly>=5.0.0

# Utilities
tqdm>=4.60.0
gitpython>=3.1.0

# Optional: For enhanced visualization
# imageio>=2.9.0
# imageio-ffmpeg>=0.4.0

# Development and Debugging
# ipython>=7.0.0
# jupyter>=1.0.0
